% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AB03.R
\name{AB03}
\alias{AB03}
\title{GMPE function for Atkinson and Boore(2003)}
\usage{
AB03(Mag, Rrup, Prd, ftype = 0, depth, Sc = 0, Sd = 0, Se = 0)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{ftype}{style of faulting.}

\item{depth}{hypocentral depth(km).}

\item{Sc}{NEHRP B Sc=0, Sd=0, Se=0, Vs>760m/s, NEHRP C Sc=1, Sd=0, Se=0, 360<Vs<760}

\item{Sd}{NEHRP D Sc=0, Sd=1, Se=0, 180<Vs<360}

\item{Se}{NEHRP E Sc=0, Sd=0, Se=1, Vs<180}
}
\value{
A list will be return, including mag, Rrup, lnY, sigma, specT, period, iflag, ftype, depth, Sc, Sd, Se.
}
\description{
\code{AB03} returns the ground-motion prediction with it sigma of Atkinson and Boore(2003) GMPE.
}
\details{
Atkinson, G., and D. Boore (2003), Empirical ground-motion relations for subduction-zone earthquakes
and their application to cascadia and other regions, Bulletin of the Seismological Society of America,
93(4), 1703-1729.
\url{http://dx.doi.org/10.1785/0120020156}
}
\examples{
AB03(6, 20, 0, 0, 10, 0, 0, 0)
AB03(7, 10, 0, 0, 10, 0, 0, 0)

}
