% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AM09.R
\name{AM09}
\alias{AM09}
\title{GMPE function for Atkinson and Macias (2009)}
\usage{
AM09(Mag, Rrup, Prd)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}
}
\value{
A list will be return, including mag, Rrup, lnY, sigma, specT, period, iflag.
}
\description{
\code{AM09} returns the ground-motion prediction with it sigma of Atkinson and Macias (2009) GMPE.
(Cascadia Subduction interface, Horizontal NEHRP B/C)
}
\details{
Atkinson, G. M., and Macias, M. (2009). Predicted Ground Motions for Great Interface Earthquakes
in the Cascadia Subduction Zone. Bulletin of the Seismological Society of America, 99(3), 1552-1578.
\url{http://doi.org/10.1785/0120080147}
}
\examples{
AM09(6, 20, 0)
AM09(7, 10, 0)

}
