% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arroyo10.R
\name{Arroyo2010}
\alias{Arroyo2010}
\title{GMPE function for  Arroyo et al. (2010)}
\usage{
Arroyo2010(Mag, Rrup, Prd)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}
}
\value{
A list will be return, including mag, Rrup, lnY, specT, iflag, period, phi1, tau1, sigma.
}
\description{
\code{Arroyo2010} returns the ground-motion prediction with it sigma of Arroyo et al. (2010) GMPE.
}
\details{
Arroyo, D. Garcia, M. Ordaz, M. A. Mora, and S. K. Singh. Strong ground-motion relations for Mexican
interplate earthquakes. Journal of Seismology, 14(4):769-785, Oct 2010.
\url{https://link.springer.com/article/10.1007/s10950-010-9200-0}
}
\examples{
Arroyo2010(6, 20, 0)
Arroyo2010(7, 10, 0)

}
