% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASB13.R
\name{ASB13}
\alias{ASB13}
\title{GMPE function for Akkar et al.(2013)}
\usage{
ASB13(Mag, Rjb, Prd = 0, ftype = 0, Vs30 = 760)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rjb}{Joyner and Boore distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{ftype}{style of faulting.}

\item{Vs30}{Vs30(m/s).}
}
\value{
A list will be return, including mag, Rjb, specT, period, lnY, sigma, iflag, ftype, Vs30, phi, tau.
}
\description{
\code{ASB13} returns the ground-motion prediction with it sigma of Akkar et al.(2013) GMPE.
}
\details{
Akkar, S., M. a. Sandikkaya, and J. J. Bommer(2013), Empirical ground-motion models for
point- and extended-source crustal earthquake scenarios in europe and the middle east,
Bulletin of Earthquake Engineering, 1-29.
\url{http://dx.doi.org/10.1007/s10518-013-9461-4}
}
\examples{
ASB13(6, 20, 0, 0, 760)
ASB13(7, 10, 0, 0, 760)

}

