% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASK14.R
\name{ASK14}
\alias{ASK14}
\title{GMPE function for Abrahamson et al.(2014) NGA-West2 model}
\usage{
ASK14(Mag, Rrup, Rjb, Prd, Dip, ftype = 0, rupwidth, Vs30 = 760,
  hwflag = 0, Ztor, Vs30_class = 1, Z1.0, Rx, Ry0, regionflag = 0,
  msasflag = 0)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Rjb}{Joyner and Boore distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{Dip}{Dip angle of the fault plane.}

\item{ftype}{style of faulting.}

\item{rupwidth}{Down-dip rupture width (km).}

\item{Vs30}{Vs30(m/s).}

\item{hwflag}{hanging-wall flag, 1 for hanging-wall.}

\item{Ztor}{Depth to the top of the finite rupture model (km).}

\item{Vs30_class}{1 measured, 0 estimated.}

\item{Z1.0}{the depth to the shear-wave velocity horizon of 1.0(km/s).}

\item{Rx}{Horizontal distance(km) from top edge of rupture. Measured perpendicular to the fault strike.}

\item{Ry0}{Horizontal distance(km) off the end of the rupture measured parallel to strike.}

\item{regionflag}{Regional attenuation flag, 0 = Global, 1 = Taiwan, 2 = China, 3 = Japan.}

\item{msasflag}{Mainshock and Aftershocks flag, 0 = Mainshocks, 1 = Aftershocks.}
}
\value{
A list will be return, including mag, dip, ftype, rupwidth, Rrup, Rjb, Vs30, hwflag,
lnY, sigma, specT, period, Ztor, iflag, vs30_class, z10, Rx, Ry0, regionflag, msasflag, phi, tau.
}
\description{
\code{ASK14} returns the ground-motion prediction with it sigma of Abrahamson et al.(2014) GMPE.
}
\details{
Abrahamson, N. A., W. J. Silva, and R. Kamai (2014), Summary of the ASK14 ground motion relation
for active crustal regions, Earthquake Spectra, 30(3), 1025-1055.
\url{http://dx.doi.org/10.1193/070913EQS198M}
}
\examples{
ASK14(6, 20, 20, 0, 90, 0, 10, 760, 0, 0, 1, 0.5, 20, 10, 0, 0)
ASK14(7, 20, 20, 0, 90, 0, 10, 760, 0, 0, 1, 0.5, 20, 10, 0, 0)

}
