% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atten2.R
\name{atten2}
\alias{atten2}
\title{Wrap function for all GMPE}
\usage{
atten2(GMPEname, Periods, Mag, Rrup, Ztor = 0, RV = 0, depth,
  ftype = 0, plot = F)
}
\arguments{
\item{GMPEname}{Name of GMPE.}

\item{Periods}{Periods of response spetra}

\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Ztor}{Depth to the top of the finite rupture model (km).}

\item{RV}{Indicator of reverse fault.}

\item{depth}{Focal depth.}

\item{ftype}{style of fault.}

\item{plot}{flag for plot or not.}
}
\value{
A list will be return, including mag, Rjb, ftype, specT, period, lnY, sigma, iflag, Vs30, phi, tau.
}
\description{
\code{atten2} returns the ground-motion prediction with it sigma of any GMPE.
}
\examples{
atten2("LL08Rock", 0, Mag=c(6,7,8), 20, 0, 0, 10, 0, FALSE)
atten2("LL08Rock", Periods=c(0.01, 0.1, 0.2, 0.5, 1, 2, 5), 6, 10, 0, 0, 10, 0, FALSE)

}
