% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI14.R
\name{BI14}
\alias{BI14}
\title{GMPE function for Bindi et al(2014)}
\usage{
BI14(Mag, Rjb, Prd = 0, ftype = 0, Vs30 = 760)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rjb}{Joyner and Boore distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{ftype}{style of faulting.}

\item{Vs30}{Vs30(m/s).}
}
\value{
A list will be return, including mag, Rjb, ftype, specT, period, lnY, sigma, iflag, Vs30, phi, tau.
}
\description{
\code{BI14} returns the ground-motion prediction with it sigma of Bindi et al.(2014) GMPE.
}
\details{
Bindi, D., M. Massa, L. Luzi, G. Ameri, F. Pacor, R. Puglia, and P. Augliera (2014),
Pan-european ground-motion prediction equations for the average horizontal component
of pGA, pGV, and 5 %-damped pSA at spectral periods up to 3.0 s using the RESORCE
dataset, Bulletin of Earthquake Engineering, 12(1), 391-430.
\url{http://dx.doi.org/10.1007/s10518-013-9525-5}
}
\examples{
BI14(6, 20, 0, 0, 760)
BI14(7, 10, 0, 0, 760)

}
