% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSSA14_tw_C01.R
\name{BSSA14.tw.C01}
\alias{BSSA14.tw.C01}
\title{GMPE function for Boore et al.(2014) NGA-West2 model}
\usage{
BSSA14.tw.C01(Mag, Rjb, Prd, Vs30, ftype = 0, Z1.0, regionflag = 0,
  basinflag = 0)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rjb}{Joyner and Boore distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{Vs30}{Vs30(m/s).}

\item{ftype}{style of faulting.}

\item{Z1.0}{the depth to the shear-wave velocity horizon of 1.0(km/s).}

\item{regionflag}{Regional attenuation flag, 0 = Global, 1 = China-Turkey, 2 = Italy-Japan.}

\item{basinflag}{Basin Adjustments.}
}
\value{
A list will be return, including mag, Rjb, specT, period, lnY, sigma, iflag, Vs30, ftype, pga4nl,
Z10, regionflag, basinflag, phi, tau.
}
\description{
\code{BSSA14.tw.C01} returns the ground-motion prediction with it sigma of Boore et al.(2014) GMPE
adjusted to Taiwan.
}
\details{
Boore, D. M., J. P. Stewart, E. Seyhan, and G. M. Atkinson (2014), NGA-west2 equations for predicting
PGA, PGV, and 5% damped PSA for shallow crustal earthquakes, Earthquake Spectra, 30(3), 1057-1085.
\url{http://dx.doi.org/10.1193/070113EQS184M}
}
\examples{
BSSA14.tw.C01(6, 20, 0, 760, 0, 0.5, 0, 0)
BSSA14.tw.C01(7, 20, 0, 760, 0, 0.5, 0, 0)

}
