% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Camp03H.R
\name{Camp03H}
\alias{Camp03H}
\title{GMPE function for Campbell and Bozorgnia(2003)}
\usage{
Camp03H(Mag, Rjb, Rseis, Prd, Svfs, Ssr, Sfr, Frv, Fth, hwflag = 0)
}
\arguments{
\item{Mag}{Earthquake local magnitude, Numeric.}

\item{Rjb}{Rupture distance(km), Numeric.}

\item{Rseis}{Seismogenic distance(km)}

\item{Prd}{Period of spectral acceleration.}

\item{Svfs}{Svfs}

\item{Ssr}{Ssr}

\item{Sfr}{Sfr}

\item{Frv}{Frv}

\item{Fth}{Fth}

\item{hwflag}{hanging-wall flag, 1 for hanging-wall.}
}
\value{
A list will be return, including mag, Rseis, Rjb, lnY, sigma, specT, period, Svfs, Ssr, Sfr, Frv, Fth, hwflag, iflag.
}
\description{
\code{Camp03H} returns the ground-motion prediction with it sigma of Campbell and Bozorgnia(2003) GMPE.
}
\details{
Campbell, Kenneth W. and Bozorgnia, Yousef (2003), Updated Near-Source Ground-Motion (Attenuation)
Relations for the Horizontal and Vertical Components of Peak Ground Acceleration and
Acceleration Response Spectra, Bulletin of the Seismological Society of America, 93(1), 314-331.
\url{http://dx.doi.org/10.1785/0120020029}
}
\examples{
Camp03H(6, 20, 20, 0, 0, 0, 0, 0, 0)
Camp03H(7, 10, 20, 0, 0, 0, 0, 0, 0)

}
