% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB14_tw_E05.R
\name{CB14.tw.E05}
\alias{CB14.tw.E05}
\title{GMPE function for adjusted Campbell and Bozorgnia(2014) NGA-West2 model}
\usage{
CB14.tw.E05(Mag, Rrup, Rjb, ftype = 0, Prd, Vs30 = 760, Ztor, Z2.5,
  Dip, depth, hwflag = 0, Rx, rupwidth, regionflag = 0)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Rjb}{Joyner-Boore distance(km), Numeric.}

\item{ftype}{style of faulting.}

\item{Prd}{Period of spectral acceleration.}

\item{Vs30}{Vs30(m/s).}

\item{Ztor}{Depth to the top of the finite rupture model (km).}

\item{Z2.5}{the depth to the shear-wave velocity horizon of 2.5(km/s).}

\item{Dip}{Dip angle of the fault plane.}

\item{depth}{hypocentral depth(km).}

\item{hwflag}{hanging-wall flag, 1 for hanging-wall.}

\item{Rx}{Horizontal distance(km) from top edge of rupture. Measured perpendicular to the fault strike.}

\item{rupwidth}{Down-dip rupture width (km).}

\item{regionflag}{Regional attenuation flag, 0 = California, 1 = Japan 2 = China, 3 = Italy.}
}
\value{
A list will be return, including mag, Rrup, Rjb, ftype, specT, period, lnY, sigma,
iflag, Vs30, Ztor, Z25, dip, depth, hwflag, Rx, rupwidth, regionflag, phi, tau.
}
\description{
\code{CB14.tw.E05} returns the ground-motion prediction with it sigma of Campbell and Bozorgnia(2014) GMPE
adjusted to Taiwan..
}
\details{
Campbell, K. W., and Y. Bozorgnia (2014), NGA-west2 ground motion model for the average
horizontal components of PGA, PGV, and 5% damped linear acceleration response spectra,
Earthquake Spectra, 30(3), 1087-1115.
\url{http://dx.doi.org/10.1193/062913EQS175M}
}
\examples{
CB14.tw.E05(6, 20, 20, 0, 0, 760, 0, 1.5, 90, 10, 0, 20, 10, 0)
CB14.tw.E05(7, 20, 20, 0, 0, 760, 0, 1.5, 90, 10, 0, 20, 10, 0)

}
