% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chao2018.R
\name{Chao2018}
\alias{Chao2018}
\title{GMPE function for Chao et al.(2018)}
\usage{
Chao2018(Mag, Rrup, Prd = 0, ftype = 0, Vs30 = 760, Vs30_class = 0,
  Ztor = 0, Z1.0 = 0.0058959, sourcetype = 0, msasflag = 0)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{ftype}{style of faulting, -1 for Normal, 0 for Strike-Slip, 1 for Reverse for sourcetype = 0.
0 for interface, 1 for intraslab for soucetype = 1.}

\item{Vs30}{Vs30(m/s).}

\item{Vs30_class}{Vs30 data class, 0 for estimated, 1 for measured.}

\item{Ztor}{Depth to the top of the finite rupture model (km).}

\item{Z1.0}{the depth to the shear-wave velocity horizon of 1.0(km/s).}

\item{sourcetype}{source type, 0 for crustal, 1 for Subduction.}

\item{msasflag}{Mainshock and Aftershocks flag, 0 = Mainshocks, 1 = Aftershocks.}
}
\value{
A list will be return, including mag, Rrup, ftype, lnY, sigma, sclass, specT, attenName, period, iflag, sourcetype, depth, phi, tau.
}
\description{
\code{Chao2018} returns the ground-motion prediction with it sigma of Chao et al.(2018) GMPE.
}
\details{
Chao et al.(2018)
}
\examples{
Chao2018(6, 20, 0, 0, 760, 0, 0, 0, 0)
Chao2018(7, 10, 0, 0, 760, 0, 0, 0, 0)

}
