% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.AGA16.tw.C01}
\alias{cnames.AGA16.tw.C01}
\title{Variable name of AGA16.tw.C01 GMPEs.}
\format{Vectors of AGA16.tw.C01 GMPE.}
\usage{
cnames.AGA16.tw.C01
}
\description{
Variable name of AGA16.tw.C01 GMPEs.
}
\keyword{datasets}
