% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.AGA16.tw.F10}
\alias{cnames.AGA16.tw.F10}
\title{Variable name of AGA16.tw.F10 GMPEs.}
\format{Vectors of AGA16.tw.F10 GMPE.}
\usage{
cnames.AGA16.tw.F10
}
\description{
Variable name of AGA16.tw.F10 GMPEs.
}
\keyword{datasets}
