% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Arroyo2010}
\alias{cnames.Arroyo2010}
\title{Variable name of Arroyo2010 GMPEs.}
\format{Vectors of Arroyo2010 GMPE.}
\usage{
cnames.Arroyo2010
}
\description{
Variable name of Arroyo2010 GMPEs.
}
\keyword{datasets}
