% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.ASK14.tw.E03}
\alias{cnames.ASK14.tw.E03}
\title{Variable name of ASK14.tw.E03 GMPEs.}
\format{Vectors of ASK14.tw.E03 GMPE.}
\usage{
cnames.ASK14.tw.E03
}
\description{
Variable name of ASK14.tw.E03 GMPEs.
}
\keyword{datasets}
