% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.BCHydroSub2018Global}
\alias{cnames.BCHydroSub2018Global}
\title{Variable name of BCHydroSub2018Global GMPEs.}
\format{
Vectors of BCHydroSub2018Global GMPE.
}
\usage{
cnames.BCHydroSub2018Global
}
\description{
Variable name of BCHydroSub2018Global GMPEs.
}
\keyword{datasets}
