% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.BCHydroSub2018SouthAM}
\alias{cnames.BCHydroSub2018SouthAM}
\title{Variable name of BCHydroSub2018SouthAM GMPEs.}
\format{
Vectors of BCHydroSub2018SouthAM GMPE.
}
\usage{
cnames.BCHydroSub2018SouthAM
}
\description{
Variable name of BCHydroSub2018SouthAM GMPEs.
}
\keyword{datasets}
