% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.BCHydroSub2018TW}
\alias{cnames.BCHydroSub2018TW}
\title{Variable name of BCHydroSub2018TW GMPEs.}
\format{
Vectors of BCHydroSub2018TW GMPE.
}
\usage{
cnames.BCHydroSub2018TW
}
\description{
Variable name of BCHydroSub2018TW GMPEs.
}
\keyword{datasets}
