% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.BCHydroSubV3}
\alias{cnames.BCHydroSubV3}
\title{Variable name of BCHydroSubV3 GMPEs.}
\format{Vectors of BCHydroSubV3 GMPE.}
\usage{
cnames.BCHydroSubV3
}
\description{
Variable name of BCHydroSubV3 GMPEs.
}
\keyword{datasets}
