% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.BI14.tw.E02}
\alias{cnames.BI14.tw.E02}
\title{Variable name of BI14.tw.E02 GMPEs.}
\format{Vectors of BI14.tw.E02 GMPE.}
\usage{
cnames.BI14.tw.E02
}
\description{
Variable name of BI14.tw.E02 GMPEs.
}
\keyword{datasets}
