% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.BSSA14}
\alias{cnames.BSSA14}
\title{Variable name of BSSA14 GMPEs.}
\format{Vectors of BSSA14 GMPE.}
\usage{
cnames.BSSA14
}
\description{
Variable name of BSSA14 GMPEs.
}
\keyword{datasets}
