% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.BSSA14.tw.B01}
\alias{cnames.BSSA14.tw.B01}
\title{Variable name of BSSA14.tw.B01 GMPEs.}
\format{Vectors of BSSA14.tw.B01 GMPE.}
\usage{
cnames.BSSA14.tw.B01
}
\description{
Variable name of BSSA14.tw.B01 GMPEs.
}
\keyword{datasets}
