% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.BSSA14.tw.E02}
\alias{cnames.BSSA14.tw.E02}
\title{Variable name of BSSA14.tw.E02 GMPEs.}
\format{Vectors of BSSA14.tw.E02 GMPE.}
\usage{
cnames.BSSA14.tw.E02
}
\description{
Variable name of BSSA14.tw.E02 GMPEs.
}
\keyword{datasets}
