% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Camp03H}
\alias{cnames.Camp03H}
\title{Variable name of Camp03H GMPEs.}
\format{Vectors of Camp03H GMPE.}
\usage{
cnames.Camp03H
}
\description{
Variable name of Camp03H GMPEs.
}
\keyword{datasets}
