% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.CB14.tw.E05}
\alias{cnames.CB14.tw.E05}
\title{Variable name of CB14.tw.E05 GMPEs.}
\format{Vectors of CB14.tw.E05 GMPE.}
\usage{
cnames.CB14.tw.E05
}
\description{
Variable name of CB14.tw.E05 GMPEs.
}
\keyword{datasets}
