% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Chao2017}
\alias{cnames.Chao2017}
\title{Variable name of Chao2017 GMPEs.}
\format{Vectors of Chao2017 GMPE.}
\usage{
cnames.Chao2017
}
\description{
Variable name of Chao2017 GMPEs.
}
\keyword{datasets}
