% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Chao2018}
\alias{cnames.Chao2018}
\title{Variable name of Chao2018 GMPEs.}
\format{Vectors of Chao2018 GMPE.}
\usage{
cnames.Chao2018
}
\description{
Variable name of Chao2018 GMPEs.
}
\keyword{datasets}
