% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Cru.Com.002}
\alias{cnames.Cru.Com.002}
\title{Variable name of Cru.Com.002 GMPEs.}
\format{Vectors of Cru.Com.002 GMPE.}
\usage{
cnames.Cru.Com.002
}
\description{
Variable name of Cru.Com.002 GMPEs.
}
\keyword{datasets}
