% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Cru.Com.003}
\alias{cnames.Cru.Com.003}
\title{Variable name of Cru.Com.003 GMPEs.}
\format{Vectors of Cru.Com.003 GMPE.}
\usage{
cnames.Cru.Com.003
}
\description{
Variable name of Cru.Com.003 GMPEs.
}
\keyword{datasets}
