% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Cru.Com.004}
\alias{cnames.Cru.Com.004}
\title{Variable name of Cru.Com.004 GMPEs.}
\format{Vectors of Cru.Com.004 GMPE.}
\usage{
cnames.Cru.Com.004
}
\description{
Variable name of Cru.Com.004 GMPEs.
}
\keyword{datasets}
