% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Cru.Com.008}
\alias{cnames.Cru.Com.008}
\title{Variable name of Cru.Com.008 GMPEs.}
\format{Vectors of Cru.Com.008 GMPE.}
\usage{
cnames.Cru.Com.008
}
\description{
Variable name of Cru.Com.008 GMPEs.
}
\keyword{datasets}
