% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Cru.Com.012}
\alias{cnames.Cru.Com.012}
\title{Variable name of Cru.Com.012 GMPEs.}
\format{Vectors of Cru.Com.012 GMPE.}
\usage{
cnames.Cru.Com.012
}
\description{
Variable name of Cru.Com.012 GMPEs.
}
\keyword{datasets}
