% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Cru.Com.013}
\alias{cnames.Cru.Com.013}
\title{Variable name of Cru.Com.013 GMPEs.}
\format{Vectors of Cru.Com.013 GMPE.}
\usage{
cnames.Cru.Com.013
}
\description{
Variable name of Cru.Com.013 GMPEs.
}
\keyword{datasets}
