% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Cru.Com.014}
\alias{cnames.Cru.Com.014}
\title{Variable name of Cru.Com.014 GMPEs.}
\format{
Vectors of Cru.Com.014 GMPE.
}
\usage{
cnames.Cru.Com.014
}
\description{
Variable name of Cru.Com.014 GMPEs.
}
\keyword{datasets}
