% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Cru.Com.015}
\alias{cnames.Cru.Com.015}
\title{Variable name of Cru.Com.015 GMPEs.}
\format{Vectors of Cru.Com.015 GMPE.}
\usage{
cnames.Cru.Com.015
}
\description{
Variable name of Cru.Com.015 GMPEs.
}
\keyword{datasets}
