% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Cru.Com.016}
\alias{cnames.Cru.Com.016}
\title{Variable name of Cru.Com.016 GMPEs.}
\format{
Vectors of Cru.Com.016 GMPE.
}
\usage{
cnames.Cru.Com.016
}
\description{
Variable name of Cru.Com.016 GMPEs.
}
\keyword{datasets}
