% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Cru.Com.017}
\alias{cnames.Cru.Com.017}
\title{Variable name of Cru.Com.017 GMPEs.}
\format{Vectors of Cru.Com.017 GMPE.}
\usage{
cnames.Cru.Com.017
}
\description{
Variable name of Cru.Com.017 GMPEs.
}
\keyword{datasets}
