% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.CY14.tw.B01}
\alias{cnames.CY14.tw.B01}
\title{Variable name of CY14.tw.B0 GMPEs.}
\format{
Vectors of CY14.tw.B0 GMPE.
}
\usage{
cnames.CY14.tw.B01
}
\description{
Variable name of CY14.tw.B0 GMPEs.
}
\keyword{datasets}
