% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.CY14.tw.E04}
\alias{cnames.CY14.tw.E04}
\title{Variable name of CY14.tw.E04 GMPEs.}
\format{Vectors of CY14.tw.E04 GMPE.}
\usage{
cnames.CY14.tw.E04
}
\description{
Variable name of CY14.tw.E04 GMPEs.
}
\keyword{datasets}
