% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Garcia05}
\alias{cnames.Garcia05}
\title{Variable name of Garcia05 GMPEs.}
\format{Vectors of Garcia05 GMPE.}
\usage{
cnames.Garcia05
}
\description{
Variable name of Garcia05 GMPEs.
}
\keyword{datasets}
