% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Gregor06}
\alias{cnames.Gregor06}
\title{Variable name of Gregor06 GMPEs.}
\format{Vectors of Gregor06 GMPE.}
\usage{
cnames.Gregor06
}
\description{
Variable name of Gregor06 GMPEs.
}
\keyword{datasets}
