% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.I14.tw.E04}
\alias{cnames.I14.tw.E04}
\title{Variable name of I14.tw.E04 GMPEs.}
\format{
Vectors of I14.tw.E04 GMPE.
}
\usage{
cnames.I14.tw.E04
}
\description{
Variable name of I14.tw.E04 GMPEs.
}
\keyword{datasets}
