% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.KAAH15}
\alias{cnames.KAAH15}
\title{Variable name of KAAH15 GMPEs.}
\format{Vectors of KAAH15 GMPE.}
\usage{
cnames.KAAH15
}
\description{
Variable name of KAAH15 GMPEs.
}
\keyword{datasets}
