% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Kanno06}
\alias{cnames.Kanno06}
\title{Variable name of Kanno06 GMPEs.}
\format{Vectors of Kanno06 GMPE.}
\usage{
cnames.Kanno06
}
\description{
Variable name of Kanno06 GMPEs.
}
\keyword{datasets}
