% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Lin11Rock}
\alias{cnames.Lin11Rock}
\title{Variable name of Lin11Rock GMPEs.}
\format{
Vectors of Lin11Rock GMPE.
}
\usage{
cnames.Lin11Rock
}
\description{
Variable name of Lin11Rock GMPEs.
}
\keyword{datasets}
