% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Lin11Soil}
\alias{cnames.Lin11Soil}
\title{Variable name of Lin11Soil GMPEs.}
\format{
Vectors of Lin11Soil GMPE.
}
\usage{
cnames.Lin11Soil
}
\description{
Variable name of Lin11Soil GMPEs.
}
\keyword{datasets}
