% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.LL08.C02}
\alias{cnames.LL08.C02}
\title{Variable name of LL08.C02 GMPEs.}
\format{Vectors of LL08.C02 GMPE.}
\usage{
cnames.LL08.C02
}
\description{
Variable name of LL08.C02 GMPEs.
}
\keyword{datasets}
