% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.LL08.E04}
\alias{cnames.LL08.E04}
\title{Variable name of LL08.E04 GMPEs.}
\format{Vectors of LL08.E04 GMPE.}
\usage{
cnames.LL08.E04
}
\description{
Variable name of LL08.E04 GMPEs.
}
\keyword{datasets}
