% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.LL08Rock}
\alias{cnames.LL08Rock}
\title{Variable name of LL08Rock GMPEs.}
\format{Vectors of LL08Rock GMPE.}
\usage{
cnames.LL08Rock
}
\description{
Variable name of LL08Rock GMPEs.
}
\keyword{datasets}
