% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.LL08Soil}
\alias{cnames.LL08Soil}
\title{Variable name of LL08Soil GMPEs.}
\format{Vectors of LL08Soil GMPE.}
\usage{
cnames.LL08Soil
}
\description{
Variable name of LL08Soil GMPEs.
}
\keyword{datasets}
