% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Mcverry2006Sub}
\alias{cnames.Mcverry2006Sub}
\title{Variable name of Mcverry2006Sub GMPEs.}
\format{
Vectors of Mcverry2006Sub GMPE.
}
\usage{
cnames.Mcverry2006Sub
}
\description{
Variable name of Mcverry2006Sub GMPEs.
}
\keyword{datasets}
