% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.NCREE2011}
\alias{cnames.NCREE2011}
\title{Variable name of NCREE2011 GMPEs.}
\format{Vectors of NCREE2011 GMPE.}
\usage{
cnames.NCREE2011
}
\description{
Variable name of NCREE2011 GMPEs.
}
\keyword{datasets}
