% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.PhungCru17}
\alias{cnames.PhungCru17}
\title{Variable name of PhungCru17 GMPEs.}
\format{
Vectors of PhungCru17 GMPE.
}
\usage{
cnames.PhungCru17
}
\description{
Variable name of PhungCru17 GMPEs.
}
\keyword{datasets}
