% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.PhungCru18}
\alias{cnames.PhungCru18}
\title{Variable name of PhungCru18 GMPEs.}
\format{Vectors of PhungCru18 GMPE.}
\usage{
cnames.PhungCru18
}
\description{
Variable name of PhungCru18 GMPEs.
}
\keyword{datasets}
