% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.PhungSub17}
\alias{cnames.PhungSub17}
\title{Variable name of PhungSub17 GMPEs.}
\format{Vectors of PhungSub17 GMPE.}
\usage{
cnames.PhungSub17
}
\description{
Variable name of PhungSub17 GMPEs.
}
\keyword{datasets}
