% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.PhungSub18}
\alias{cnames.PhungSub18}
\title{Variable name of PhungSub18 GMPEs.}
\format{Vectors of PhungSub18 GMPE.}
\usage{
cnames.PhungSub18
}
\description{
Variable name of PhungSub18 GMPEs.
}
\keyword{datasets}
