% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Sinotech12}
\alias{cnames.Sinotech12}
\title{Variable name of Sinotech12 GMPEs.}
\format{Vectors of Sinotech12 GMPE.}
\usage{
cnames.Sinotech12
}
\description{
Variable name of Sinotech12 GMPEs.
}
\keyword{datasets}
