% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Sub.Inter.Com.006}
\alias{cnames.Sub.Inter.Com.006}
\title{Variable name of Sub.Inter.Com.006 GMPEs.}
\format{Vectors of Sub.Inter.Com.006 GMPE.}
\usage{
cnames.Sub.Inter.Com.006
}
\description{
Variable name of Sub.Inter.Com.006 GMPEs.
}
\keyword{datasets}
