% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Sub.Inter.Com.007}
\alias{cnames.Sub.Inter.Com.007}
\title{Variable name of Sub.Inter.Com.007 GMPEs.}
\format{Vectors of Sub.Inter.Com.007 GMPE.}
\usage{
cnames.Sub.Inter.Com.007
}
\description{
Variable name of Sub.Inter.Com.007 GMPEs.
}
\keyword{datasets}
