% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Sub.Inter.Com.009}
\alias{cnames.Sub.Inter.Com.009}
\title{Variable name of Sub.Inter.Com.009 GMPEs.}
\format{Vectors of Sub.Inter.Com.009 GMPE.}
\usage{
cnames.Sub.Inter.Com.009
}
\description{
Variable name of Sub.Inter.Com.009 GMPEs.
}
\keyword{datasets}
