% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Sub.Inter.Com.013}
\alias{cnames.Sub.Inter.Com.013}
\title{Variable name of Sub.Inter.Com.013 GMPEs.}
\format{Vectors of Sub.Inter.Com.013 GMPE.}
\usage{
cnames.Sub.Inter.Com.013
}
\description{
Variable name of Sub.Inter.Com.013 GMPEs.
}
\keyword{datasets}
