% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Sub.Inter.Com.016}
\alias{cnames.Sub.Inter.Com.016}
\title{Variable name of Sub.Inter.Com.016 GMPEs.}
\format{Vectors of Sub.Inter.Com.016 GMPE.}
\usage{
cnames.Sub.Inter.Com.016
}
\description{
Variable name of Sub.Inter.Com.016 GMPEs.
}
\keyword{datasets}
