% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Sub.Intra.Com.001}
\alias{cnames.Sub.Intra.Com.001}
\title{Variable name of Sub.Intra.Com.001 GMPEs.}
\format{Vectors of Sub.Intra.Com.001 GMPE.}
\usage{
cnames.Sub.Intra.Com.001
}
\description{
Variable name of Sub.Intra.Com.001 GMPEs.
}
\keyword{datasets}
