% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Sub.Intra.Com.005}
\alias{cnames.Sub.Intra.Com.005}
\title{Variable name of Sub.Intra.Com.005 GMPEs.}
\format{Vectors of Sub.Intra.Com.005 GMPE.}
\usage{
cnames.Sub.Intra.Com.005
}
\description{
Variable name of Sub.Intra.Com.005 GMPEs.
}
\keyword{datasets}
