% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Sub.Intra.Com.011}
\alias{cnames.Sub.Intra.Com.011}
\title{Variable name of Sub.Intra.Com.011 GMPEs.}
\format{Vectors of Sub.Intra.Com.011 GMPE.}
\usage{
cnames.Sub.Intra.Com.011
}
\description{
Variable name of Sub.Intra.Com.011 GMPEs.
}
\keyword{datasets}
