% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Youngs97Rock}
\alias{cnames.Youngs97Rock}
\title{Variable name of Youngs97Rock GMPEs.}
\format{
Vectors of Youngs97Rock GMPE.
}
\usage{
cnames.Youngs97Rock
}
\description{
Variable name of Youngs97Rock GMPEs.
}
\keyword{datasets}
