% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Zh16Cru}
\alias{cnames.Zh16Cru}
\title{Variable name of Zh16Cru GMPEs.}
\format{
Vectors of Zh16Cru GMPE.
}
\usage{
cnames.Zh16Cru
}
\description{
Variable name of Zh16Cru GMPEs.
}
\keyword{datasets}
