% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Zh16Inter}
\alias{cnames.Zh16Inter}
\title{Variable name of Zh16Inter GMPEs.}
\format{Vectors of Zh16Inter GMPE.}
\usage{
cnames.Zh16Inter
}
\description{
Variable name of Zh16Inter GMPEs.
}
\keyword{datasets}
