% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Zh16Slab}
\alias{cnames.Zh16Slab}
\title{Variable name of Zh16Slab GMPEs.}
\format{Vectors of Zh16Slab GMPE.}
\usage{
cnames.Zh16Slab
}
\description{
Variable name of Zh16Slab GMPEs.
}
\keyword{datasets}
