% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnames.Zh16Sub}
\alias{cnames.Zh16Sub}
\title{Variable name of Zh16Sub GMPEs.}
\format{
Vectors of Zh16Sub GMPE.
}
\usage{
cnames.Zh16Sub
}
\description{
Variable name of Zh16Sub GMPEs.
}
\keyword{datasets}
