% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crustal_Common_Form.R
\name{Cru.Com.017}
\alias{Cru.Com.017}
\title{GMPE function for Crustal Common form 017 (2017)}
\usage{
Cru.Com.017(Mag, Rrup, Rjb, Ztor, ftype = 0, Dip, rupwidth, Rx, hwflag = 0,
  Prd)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Rjb}{Joyner and Boore distance(km), Numeric.}

\item{Ztor}{Depth to the top of the finite rupture model (km).}

\item{ftype}{style of faulting.}

\item{Dip}{Dip angle of the fault plane.}

\item{rupwidth}{Down-dip rupture width (km).}

\item{Rx}{Horizontal distance(km) from top edge of rupture. Measured perpendicular to the fault strike.}

\item{hwflag}{hanging-wall flag, 1 for hanging-wall.}

\item{Prd}{Period of spectral acceleration.}
}
\value{
A list will be return, including mag, Rrup, Rjb, Ztor, ftype, dip, rupwidth, Rx,
 hwflag, pecT, lnY, sigma, iflag.
}
\description{
\code{Cru.Com.017} returns the ground-motion prediction with it sigma of Crustal Common form 017 GMPE.
}
\details{
Crustal Common form 017
}
\examples{
Cru.Com.017(6, 20, 20, 5, 0, 90, 20, 20, 0, 0)
Cru.Com.017(7, 20, 20, 2, 0, 90, 20, 20, 0, 0 )

}
