% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CY08.R
\name{CY08}
\alias{CY08}
\title{GMPE function for Chiou and Youngs(2008) NGA model}
\usage{
CY08(Mag, Rrup, Rjb, Prd, Vs30, Dip, Ztor, ftype = 0, Z1.0,
  Vs30_class = 1, hwflag = 0, Rx, regionflag = 0)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Rjb}{Joyner and Boore distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{Vs30}{Vs30(m/s).}

\item{Dip}{Dip angle of the fault plane.}

\item{Ztor}{Depth to the top of the finite rupture model (km).}

\item{ftype}{style of faulting.}

\item{Z1.0}{the depth to the shear-wave velocity horizon of 1.0(km/s).}

\item{Vs30_class}{1 measured, 0 estimated.}

\item{hwflag}{hanging-wall flag, 1 for hanging-wall.}

\item{Rx}{Horizontal distance(km) from top edge of rupture. Measured perpendicular to the fault strike.}

\item{regionflag}{0 Global, 1 Japan and Italy, 2 Wenchuan (note only for M7.9)}
}
\value{
A list will be return, including mag, Rrup, Rjb, specT, period, lnY, sigma, iflag, Vs30,
dip, Ztor, ftype, Z10, vs3-_class, hwflag, Rx, regionflag, phi, tau.
}
\description{
\code{CY08} returns the ground-motion prediction with it sigma of Chiou and Youngs(2008) GMPE.
}
\details{
BrianS-J. Chiou and Robert R. Youngs (2008) An NGA Model for the Average Horizontal Component
of Peak Ground Motion and Response Spectra. Earthquake Spectra, Vol. 24, No. 1, pp. 173-215.
\url{http://dx.doi.org/10.1193/1.2894832}
}
\examples{
CY08(6, 20, 20, 0, 760, 90, 0, 0, 0.5, 1, 0, 20, 0)
CY08(7, 20, 20, 0, 760, 90, 0, 0, 0.5, 1, 0, 20, 0)

}
