% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CY14_tw_C01.R
\name{CY14.tw.C01}
\alias{CY14.tw.C01}
\title{GMPE function for Chiou and Youngs(2014) NGA-West2 model}
\usage{
CY14.tw.C01(Mag, Rrup, Rjb, Prd, Vs30, Dip, Ztor, ftype = 0, Z1.0,
  Vs30_class = 1, hwflag = 0, Rx, regionflag = 0)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Rjb}{Joyner and Boore distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{Vs30}{Vs30(m/s).}

\item{Dip}{Dip angle of the fault plane.}

\item{Ztor}{Depth to the top of the finite rupture model (km).}

\item{ftype}{style of faulting.}

\item{Z1.0}{the depth to the shear-wave velocity horizon of 1.0(km/s).}

\item{Vs30_class}{1 measured, 0 estimated.}

\item{hwflag}{hanging-wall flag, 1 for hanging-wall.}

\item{Rx}{Horizontal distance(km) from top edge of rupture. Measured perpendicular to the fault strike.}

\item{regionflag}{0 Global, 1 Japan and Italy, 2 Wenchuan (note only for M7.9)}
}
\value{
A list will be return, including mag, Rrup, Rjb, specT, period, lnY, sigma, iflag, Vs30,
dip, Ztor, ftype, Z10, vs3-_class, hwflag, Rx, regionflag, phi, tau.
}
\description{
\code{CY14.tw.C01} returns the ground-motion prediction with it sigma of Chiou and Youngs(2014) GMPE
adjusted to Taiwan.
}
\details{
Chiou, B. S.-J., and R. R. Youngs (2014), Update of the chiou and youngs NGA model for
the average horizontal component of peak ground motion and response spectra,
Earthquake Spectra, 30(3), 1117-1153.
\url{http://dx.doi.org/10.1193/072813EQS219M}
}
\examples{
CY14.tw.C01(6, 20, 20, 0, 760, 90, 0, 0, 0.5, 1, 0, 20, 0)
CY14.tw.C01(7, 20, 20, 0, 760, 90, 0, 0, 0.5, 1, 0, 20, 0)

}
