% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Garcia05.R
\name{Garcia05}
\alias{Garcia05}
\title{GMPE function for Garcia et al.(2005)}
\usage{
Garcia05(Mag, Rrup, Prd, depth)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{depth}{hypocentral depth(km).}
}
\value{
A list will be return, including mag, Rrup, specT, period, lnY, sigma, iflag, depth.
}
\description{
\code{Garcia05} returns the ground-motion prediction with it sigma of Garcia et al.(2005) GMPE.
}
\details{
Garcia, D., Singh, S. K., Herraiz, M., Ordaz, M., &Pacheco, J. F. (2005). Inslab Earthquakes of Central
Mexico: Peak Ground-Motion Parameters and Response Spectra. Bulletin of the Seismological Society of America,
95(6), 2272 LP-2282. Retrieved from \url{http://www.bssaonline.org/content/95/6/2272.abstract}
}
\examples{
Garcia05(6, 20, 0, 35)
Garcia05(7, 10, 0, 60)

}
