% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GMPE_prd.R
\name{get.GMPE.prd}
\alias{get.GMPE.prd}
\title{To obtain usable period range of GMPE}
\usage{
get.GMPE.prd(GMPE)
}
\arguments{
\item{GMPE}{Name of GMPE.}
}
\value{
A data.frame of period.min and period.max will be return.
}
\description{
\code{get.GMPE.prd} returns the the minimum and maximum period of any GMPE.
}
\examples{
get.GMPE.prd("ASK14")
get.GMPE.prd("CY14")

}
