% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GK16.R
\name{GK16}
\alias{GK16}
\title{GMPE function for Graizer and Kalkan(2016)}
\usage{
GK16(Mag, Rrup, ftype = 0, Prd, Vs30 = 760, q0 = 150, Z1.5)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{ftype}{style of faulting.}

\item{Prd}{Period of spectral acceleration.}

\item{Vs30}{Vs30(m/s).}

\item{q0}{Quality factor (Q0).}

\item{Z1.5}{Depth to the 1.5km/s shear-wave velocity isosurfac.}
}
\value{
A list will be return, including mag, Rrup, specT, ftype, period, lnY, sigma, iflag, Vs30, q0, Z15.
}
\description{
\code{GK16} returns the ground-motion prediction with it sigma of Graizer and Kalkan(2016) GMPE.
}
\details{
Graizer and E. Kalkan.(2016) Summary of the GK15 ground-motion prediction equation for
horizontal PGA and 5% damped PSA from shallow crustal continental earthquakes.
Bulletin of the Seismological Society of America, 106(2), 687-707.
\url{http://dx.doi.org/10.1785/0120150194}
}
\examples{
GK16(6, 20, 0, 0, 760, 150, 1)
GK16(7, 10, 0, 0, 760, 150, 1)

}
