% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GMPEhaz.model}
\alias{GMPEhaz.model}
\title{GMPE model list.}
\format{dataframe of GMPE model list.}
\usage{
GMPEhaz.model
}
\description{
To obtain useable period range for each GMPE model.
}
\keyword{datasets}
