% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gregor06.R
\name{Gregor06}
\alias{Gregor06}
\title{GMPE function for Gregor et al.(2006)}
\usage{
Gregor06(Mag, Rrup, Prd, Vs30)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{Vs30}{Vs30(m/s).}
}
\value{
A list will be return, including mag, Rrup, lnY, sigma, specT, attenName, Vs30, period, iflag.
}
\description{
\code{Gregor06} returns the ground-motion prediction with it sigma of Gregor et al.(2006) GMPE.
}
\details{
Gregor, N., W. Silva, I. Wong, and R. Youngs (2006), Updated response spectral attenuation relationship
for cascadia subduction zone megathrust earthquake, Seismological Research Letters, 77(2), 325-326.
}
\examples{
Gregor06(6, 20, 0, 760)
Gregor06(7, 10, 0, 760)

}
