% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KAAH15.R
\name{KAAH15}
\alias{KAAH15}
\title{GMPE function for Kale et al.(2015)}
\usage{
KAAH15(Mag, Rjb, Prd = 0, ftype = 0, Vs30 = 760, regionflag = 0)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rjb}{Joyner and Boore distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{ftype}{style of faulting.}

\item{Vs30}{Vs30(m/s).}

\item{regionflag}{0 for Turkey, 1 for Iran.}
}
\value{
A list will be return, including mag, Rjb, specT, period, lnY, sigma, iflag, ftype, Vs30, pgaref, region.
}
\description{
\code{KAAH15} returns the ground-motion prediction with it sigma of Kale et al.(2015) GMPE.
}
\details{
Kale, O., S. Akkar, A. Ansari, and H. Hamzehloo (2015), A ground-Motion predictive model for
Iran and Turkey for horizontal PGA, PGV, and 5% Damped Response Spectrum: Investigation of
Possible Regional Effects, Bulletin of the Seismological Society of America, 105(2A), 963-980.
\url{http://dx.doi.org/10.1785/0120140134}
}
\examples{
KAAH15(6, 20, 0, 0, 760, 0)
KAAH15(7, 10, 0, 0, 760, 0)

}
