% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kanno06.R
\name{Kanno06}
\alias{Kanno06}
\title{GMPE function for Kanno et al.(2006)}
\usage{
Kanno06(Mag, Rrup, Prd, Vs30, depth)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{Vs30}{Vs30(m/s).}

\item{depth}{hypocentral depth(km).}
}
\value{
A list will be return, including mag, Rrup, specT, period, lnY, sigma, iflag, Vs30, depth.
}
\description{
\code{Kanno06} returns the ground-motion prediction with it sigma of Kanno et al.(2006) GMPE.
}
\details{
Kanno, T., A. Narita, N. Morikawa, H. Fujiwara, and Y. Fukushima (2006), A new attenuation relation
for strong ground motion in japan based on recorded data, Bulletin of the Seismological Society of America,
96(3), 879-897.
\url{http://dx.doi.org/10.1785/0120050138}
}
\examples{
Kanno06(6, 20, 0, 760, 10)
Kanno06(7, 10, 0, 760, 10)

}
