% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lin09.R
\name{Lin09}
\alias{Lin09}
\title{GMPE function for Lin (2009)}
\usage{
Lin09(Mag, Rrup, Prd, Vs30 = 760, ftype)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{Vs30}{Vs30(m/s)}

\item{ftype}{style of faulting.}
}
\value{
A list will be return, including mag, Rrup, specT, period, lnY, sigma, Vs30, iflag.
}
\description{
\code{Lin09} returns the ground-motion prediction with it sigma of Lin (2009) GMPE.
}
\details{
Lin, P.-S. (2009). Ground-motion attenuation relationship and path-effect
study using Taiwan data set, Ph.D. Thesis, Nation Central University, Taiwan (in Chinese).
}
\examples{
Lin09(6, 20, 0, 760, 0)
Lin09(7, 10, 0, 760, 0)

}
