% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lin11.R
\name{Lin11}
\alias{Lin11}
\title{GMPE function for Lin et al.(2011)}
\usage{
Lin11(Mag, Rrup, Prd, hwflag = 0, Vs30 = 760)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{hwflag}{hanging-wall flag, 1 for hanging-wall.}

\item{Vs30}{Vs30(m/s)}
}
\value{
A list will be return, including mag, Rrup, specT, period, lnY, sigma, iflag.
}
\description{
\code{Lin11} returns the ground-motion prediction (rock or soil site base on Vs30)
with it sigma of Lin et al.(2011) GMPE.
}
\details{
Lin, P.-S., C.-T. Lee, C.-T. Cheng, and C.-H. Sung (2011), Response spectral attenuation relations
for shallow crustal earthquakes in taiwan, Engineering Geology, 121(3-4), 150-164.
\url{http://dx.doi.org/10.1016/j.enggeo.2011.04.019}
}
\examples{
Lin11(6, 20, 0, 0, 760)
Lin11(7, 10, 0, 0, 760)

}
