% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lin11.R
\name{Lin11Soil}
\alias{Lin11Soil}
\title{GMPE function for Lin et al.(2011)}
\usage{
Lin11Soil(Mag, Rrup, Prd, hwflag = 0)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{hwflag}{hanging-wall flag, 1 for hanging-wall.}
}
\value{
A list will be return, including mag, Rrup, specT, period, lnY, sigma, iflag.
}
\description{
\code{Lin11Soil} returns the ground-motion prediction (soil site) with it sigma of Lin et al.(2011) GMPE.
}
\details{
Lin, P.-S., C.-T. Lee, C.-T. Cheng, and C.-H. Sung (2011), Response spectral attenuation relations
for shallow crustal earthquakes in taiwan, Engineering Geology, 121(3-4), 150-164.
\url{http://dx.doi.org/10.1016/j.enggeo.2011.04.019}
}
\examples{
Lin11Soil(6, 20, 0, 0)
Lin11Soil(7, 10, 0, 0)

}
