% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL08.R
\name{LL08}
\alias{LL08}
\title{GMPE function for Lin and Lee(2008)}
\usage{
LL08(Mag, Rrup, depth, ftype, Prd, Vs30 = 760)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{depth}{hypocentral depth(km).}

\item{ftype}{fytpe=0 for interface, ftype=1 for intraslab}

\item{Prd}{Period of spectral acceleration.}

\item{Vs30}{Vs30(m/s)}
}
\value{
A list will be return, including mag, Rrup, specT, period, lnY, sigma, iflag, depth, ftype.
}
\description{
\code{LL08} returns the ground-motion prediction (Rock or Soil site base on Vs30)
with it sigma of Lin and Lee(2008) GMPE.
}
\details{
Lin, P., and C. Lee (2008), Ground-motion attenuation relationships for subduction-zone earthquakes in
northeastern taiwan, Bulletin of the Seismological Society of America, 98(1), 220-240.
\url{http://dx.doi.org/10.1785/0120060002}
}
\examples{
LL08(6, 20, 10, 0, 0, 760)
LL08(7, 10, 10, 0, 0, 760)

}
