% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBR17.R
\name{MBR17}
\alias{MBR17}
\title{GMPE function for Montalva et al.(2017)}
\usage{
MBR17(Mag, Rrup, Prd, ftype = 0, Vs30, forearc = 1, depth, Rhypo)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{ftype}{fytpe=0 for interface (use rupture distance), ftype=1 for intraslab (use hypocentral distance).}

\item{Vs30}{Vs30(m/s).}

\item{forearc}{0 = Forearc site, 1 = Backarc site.}

\item{depth}{hypocentral depth(km).}

\item{Rhypo}{hypocentral distance(km).}
}
\value{
A list will be return, including mag, ftype, Rrup, Vs30, lnY, sigma, specT, period, iflag, forearc, depth, Rhypo.
}
\description{
\code{MBR17} returns the ground-motion prediction with it sigma of Montalva et al.(2017) GMPE.
}
\details{
Montalva, G. A., Bastias, N., and Rodriguez-Marek, A. (2017). Ground-Motion Prediction Equation for the
Chilean Subduction Zone. Bulletin of the Seismological Society of America, 107(2), 901-911.
\url{http://doi.org/10.1785/0120160221}
}
\examples{
MBR17(6, 20, 0, 0, 760, 1, 10, 20)
MBR17(7, 10, 0, 0, 760, 1, 10, 20)

}
