% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McVerry06Sub.R
\name{Mcverry2006Sub}
\alias{Mcverry2006Sub}
\title{GMPE function for Mcverry et al.(2006) Subduction GMPE.}
\usage{
Mcverry2006Sub(Mag, Rrup, Prd, ftype = 0, Ztor, Dip, rupwidth, depth,
  Sc = 0, Sd = 0)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{ftype}{fytpe=0 for interface, ftype=1 for intraslab.}

\item{Ztor}{Depth to the top of the finite rupture model (km).}

\item{Dip}{Dip angle of the fault plane.}

\item{rupwidth}{Down-dip rupture width (km).}

\item{depth}{hypocentral depth(km).}

\item{Sc}{1 for site class C, 0 for others.}

\item{Sd}{1 for site class D, 0 for others.}
}
\value{
A list will be return, including mag, Rrup, Rjb, ftype, specT, period, lnY, sigma,
iflag, Vs30, Ztor, Z25, dip, depth, hwflag, Rx, rupwidth, regionflag, phi, tau.
}
\description{
\code{Mcverry2006Sub} returns the ground-motion prediction with it sigma of Mcverry et al.(2006) Subduction GMPE.
}
\details{
Mcverry, G. H., Zhao, J. X., Abrahamson, N. A., and Somerville, P. G. (2006). New Zealand acceleration response spectrum
attenuation relations for crustal and subduction zone earthquakes,
Bulletin of the New Zealand Society for Earthquake Engineering, 39(1), 1-58.
\url{http://bulletin.nzsee.org.nz/39/1/0001}
}
\examples{
Mcverry2006Sub(6, 20, 0, 0, 0, 90, 10, 10, 0, 0)
Mcverry2006Sub(7, 20, 0, 0, 0, 90, 10, 10, 0, 0)

}
