% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCREE2011.R
\name{NCREE2011}
\alias{NCREE2011}
\title{GMPE function for NCREE(2011)}
\usage{
NCREE2011(Mag, Rrup, Prd)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}
}
\value{
A list will be return, including mag, Rrup, specT, period, lnY, sigma.
}
\description{
\code{NCREE2011} returns the ground-motion prediction (rock site, Vs30>=360m/s)
with it sigma of NCREE(2011) GMPE.
}
\details{
國家地震工程研究中心(2011)，面震源與衰減律參數研究及核電廠地盤反應量測，行政院原子能委員會核能研究所委託研究報告。
}
\examples{
NCREE2011(6, 20, 0)
NCREE2011(7, 10, 0)

}
