% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhungCru17.R
\name{PhungCru17}
\alias{PhungCru17}
\title{GMPE function for Phung et al.(2017)}
\usage{
PhungCru17(Mag, Rrup, Prd, Vs30, Dip, Ztor, ftype = 0, Z1.0,
  Vs30_class = 1, msasflag = 0, regionflag = 1)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{Vs30}{Vs30(m/s).}

\item{Dip}{Dip angle of the fault plane.}

\item{Ztor}{Depth to the top of the finite rupture model (km).}

\item{ftype}{style of faulting.}

\item{Z1.0}{the depth to the shear-wave velocity horizon of 1.0(km/s).}

\item{Vs30_class}{1 measured, 0 estimated.}

\item{msasflag}{Mainshock and Aftershocks flag, 0 = Mainshocks, 1 = Aftershocks.}

\item{regionflag}{0 Global, 1 Taiwan}
}
\value{
A list will be return, including m, Rrup, specT, period2, lnY, sigma, iflag,
   vs, Delta, DTor, Ftype, depthvs10, vs30_class, regionflag, msasflag, phi, tau.
}
\description{
\code{PhungCru17} returns the ground-motion prediction with it sigma of Phung et al.(2017) GMPE.
}
\details{
Phung et al.(2017) Crustal GMPE
}
\examples{
PhungCru17(6, 20, 0, 760, 90, 0, 0, 0.5, 1, 0,  1)
PhungCru17(7, 20, 0, 760, 90, 0, 0, 0.5, 1, 0,  1)

}
