% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sinotech12.R
\name{Sinotech12}
\alias{Sinotech12}
\title{GMPE function for Sinotech(2012)}
\usage{
Sinotech12(Mag, Rrup, Prd, Vs30 = 760, ftype)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{Vs30}{Vs30(m/s)}

\item{ftype}{style of faulting.}
}
\value{
A list will be return, including mag, Rrup, specT, period, lnY, sigma, Vs30, iflag.
}
\description{
\code{Sinotech12} returns the ground-motion prediction with it sigma of Sinotech(2012) GMPE.
}
\details{
Sinotech (2012). (in Chinese).
}
\examples{
Sinotech12(6, 20, 0, 760, 0)
Sinotech12(7, 10, 0, 760, 0)

}
