% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subduction_Common_Form.R
\name{Sub.Inter.Com.004}
\alias{Sub.Inter.Com.004}
\title{GMPE function for Subduction interface common form 004 (2017)}
\usage{
Sub.Inter.Com.004(Mag, Rrup, Ztor, Prd)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Ztor}{Depth to the top of the finite rupture model (km).}

\item{Prd}{Period of spectral acceleration.}
}
\value{
A list will be return, including mag, Rrup, Ztor, pecT, lnY, sigma, iflag.
}
\description{
\code{Sub.Inter.Com.004} returns the ground-motion prediction with it sigma of Subduction interface Common form 004 GMPE.
}
\details{
Subduction interface common form 004
}
\examples{
Sub.Inter.Com.004(6, 20, 5, 0)
Sub.Inter.Com.004(7, 20, 2, 0)

}
