% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subduction_Common_Form.R
\name{Sub.Intra.Com.003}
\alias{Sub.Intra.Com.003}
\title{GMPE function for Subduction intraslab common form 003 (2017)}
\usage{
Sub.Intra.Com.003(Mag, Rrup, Ztor, Prd)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Ztor}{Depth to the top of the finite rupture model (km).}

\item{Prd}{Period of spectral acceleration.}
}
\value{
A list will be return, including mag, Rrup, Ztor, pecT, lnY, sigma, iflag.
}
\description{
\code{Sub.Intra.Com.003} returns the ground-motion prediction with it sigma of Subduction intraslab Common form 003 GMPE.
}
\details{
Subduction intraslab common form 003
}
\examples{
Sub.Intra.Com.003(6, 20, 5, 0)
Sub.Intra.Com.003(7, 20, 2, 0)

}
