% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Youngs97.R
\name{Youngs97Soil}
\alias{Youngs97Soil}
\title{GMPE function for Youngs et al.(1997)}
\usage{
Youngs97Soil(Mag, Rrup, Prd, ftype = 0, depth)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{Prd}{Period of spectral acceleration.}

\item{ftype}{fytpe=0 for interface, ftype=1 for intraslab}

\item{depth}{hypocentral depth(km).}
}
\value{
A list will be return, including mag, Rrup, lnY, sigma, period, specT, ftype, depth, iflag.
}
\description{
\code{Youngs97Soil} returns the ground-motion prediction (soil site) with it sigma of Youngs et al.(1997) GMPE.
}
\details{
Youngs, R. R., S.-J. Chiou, W. J. Silva, and J. R. Humphrey (1997), Strong ground motion attenuation
relationships for subduction zone earthquakes, Seismological Research Letters, 68(1), 58-73.
\url{http://dx.doi.org/10.1785/gssrl.68.1.58}
}
\examples{
Youngs97Soil(6, 20, 0, 0, 10)
Youngs97Soil(7, 10, 0, 0, 10)

}
