% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zh06.R
\name{Zh06}
\alias{Zh06}
\title{GMPE function for Zhao et al.(2006)}
\usage{
Zh06(Mag, Rrup, ftype = 0, sclass = 1, Prd = 0, sourceclass = 0,
  depth = 10)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{ftype}{style of faulting.}

\item{sclass}{site condition.}

\item{Prd}{Period of spectral acceleration.}

\item{sourceclass}{source type, 0 for crustal, 1 for subduction interface, 2 for subduction intraslab.}

\item{depth}{focal depth(km).}
}
\value{
A list will be return, including mag, Rrup, ftype, lnY, sigma, sclass, specT, attenName, period, iflag, sourcetype, depth, phi, tau.
}
\description{
\code{Zh06} returns the ground-motion prediction with it sigma of Zhao et al.(2006) GMPE.
}
\details{
Zhao, J. et al. (2006), Attenuation relations of strong ground motion in japan using
site classification based on predominant period,
Bulletin of the Seismological Society of America, 96(3), 898-913.
\url{http://dx.doi.org/10.1785/0120050122}
}
\examples{
Zh06(6, 20, 0, 1.0, 0, 0.0, 10)
Zh06(7, 10, 0, 1.0, 0, 0.0, 10)

}
