% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zh16Cru.R
\name{Zh16Cru}
\alias{Zh16Cru}
\title{GMPE function for Zhao et al.(2016)}
\usage{
Zh16Cru(Mag, Rrup, ftype = 0, sclass = 1, Prd = 0, sourceclass = 0,
  depth = 10)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{ftype}{style of faulting, -1 for Normal, -0.5 for Normal-Oblique, 0 for Strike-Slip, 0.5 for Reverse-Oblique and 1 for Reverse.}

\item{sclass}{site condition, 0 for Hard Rock, 1 for SC I, 2 for SC II, 3 for SC III, 4 for SC IV.}

\item{Prd}{Period of spectral acceleration.}

\item{sourceclass}{source type, 0 for crustal, 3 for Upper mantle.}

\item{depth}{focal depth(km).}
}
\value{
A list will be return, including mag, Rrup, ftype, lnY, sigma, sclass, specT, attenName, period, iflag, sourcetype, depth, phi, tau.
}
\description{
\code{Zh16Cru} returns the ground-motion prediction with it sigma of Zhao et al.(2016) GMPE.
}
\details{
Zhao, J. et al. (2016), Ground-Motion Prediction Equations for Shallow Crustal and Upper-Mantle
Earthquakes in Japan Using Site Class and Simple Geometric Attenuation Functions,
Bulletin of the Seismological Society of America, 106(4), 1552-1569.
\url{http://dx.doi.org/10.1785/0120150063}
}
\examples{
Zh16Cru(6, 20, 0, 1.0, 0, 0.0, 10)
Zh16Cru(7, 10, 0, 1.0, 0, 0.0, 10)

}
