% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zh16Sub.R
\name{Zh16Sub}
\alias{Zh16Sub}
\title{GMPE function for Zhao et al.(2016) Subduction GMPE}
\usage{
Zh16Sub(Mag, Rrup, sclass = 1, Prd = 0, sourceclass = 1,
  depth = 10)
}
\arguments{
\item{Mag}{Earthquake momnet magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}

\item{sclass}{site condition, 0 for Hard Rock, 1 for SC I, 2 for SC II, 3 for SC III, 4 for SC IV.}

\item{Prd}{Period of spectral acceleration.}

\item{sourceclass}{source type, 1 for Subduction - Interface, 2 for Subduction - Slab.}

\item{depth}{focal depth(km).}
}
\value{
A list will be return, including mag, Rrup, ftype, lnY, sigma, sclass, specT, attenName, period, iflag, sourcetype, depth, phi, tau.
}
\description{
\code{Zh16Sub} returns the ground-motion prediction with it sigma of Zhao et al.(2016) GMPE.
}
\details{
Zhao, J. et al. (2016),Ground-Motion Prediction Equations for Subduction Interface Earthquakes
in Japan Using Site Class and Simple Geometric Attenuation Functions,
Bulletin of the Seismological Society of America, 106(4), 1518-1534.
\url{http://dx.doi.org/10.1785/0120150034}
#'Zhao, J. et al. (2016),Ground-Motion Prediction Equations for Subduction Slab Earthquakes in
Japan Using Site Class and Simple Geometric Attenuation Functions,
Bulletin of the Seismological Society of America, 106(4), 1535-1551.
\url{http://dx.doi.org/10.1785/0120150056}
}
\examples{
Zh16Sub(6, 20, 1.0, 0, 1.0, 10)
Zh16Sub(7, 10, 1.0, 0, 1.0, 10)

}
