% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loh96.R
\name{Loh96}
\alias{Loh96}
\title{GMPE function for Loh et al.(1996) PGA only}
\usage{
Loh96(Mag, Rrup)
}
\arguments{
\item{Mag}{Earthquake local magnitude, Numeric.}

\item{Rrup}{Rupture distance(km), Numeric.}
}
\value{
A list will be return, including mag, Rrup, lnY, sigma, period.
}
\description{
\code{Loh96} returns the ground-motion prediction (rock site) with it sigma of Loh et al.(1996) GMPE.
}
\examples{
Loh96(6, 20)
Loh96(7, 10)

}
